Update README.md
### Spotify Clone Code Project

The Spotify Clone Code Project is a front-end web application designed to replicate the look and feel of Spotify. Using only HTML and CSS, this project focuses on creating a visually appealing and responsive design that mirrors the core interface of Spotify. Key features include a modern user interface, dynamic layout, and styling elements that showcase album artwork, playlists, and user controls. This project is perfect for developers looking to hone their skills in HTML and CSS by building a sophisticated, real-world application interface.